pyproject.toml
[project]
name = "role-based-fastapi"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "alembic>=1.17.2",
    "bcrypt>=4.0.0",
    "fastapi>=0.121.2",
    "passlib[bcrypt]>=1.7.4",
    "psycopg2>=2.9.11",
    "psycopg2-binary>=2.9.11",
    "pydantic[email]>=2.12.4",
    "python-dotenv>=1.2.1",
    "python-jose[cryptography]>=3.5.0",
    "sqlalchemy>=2.0.44",
    "uvicorn>=0.38.0",
]
